+++ client/src/components/Partials/HomePage.jsx
@@ -4,13 +4,21 @@ import { Link } from 'react-router-dom';
 import axios from 'axios';
 import { backendURL } from '../../config/config';
 import SuccessMessage from './Messages/SuccessMessage';
+import Cookies from 'js-cookie';
 
 export default class HomePage extends Component {
   componentDidMount() {
-    document.title = "Home - SnailyCAD"
+    document.title = 'Home - SnailyCAD';
   }
   render() {
-    return getSession() ? (<div><LoggedInSection /><Credits /></div>) : <NotLoggedInSection />;
+    return getSession() ? (
+      <div>
+        <LoggedInSection />
+        <Credits />
+      </div>
+    ) : (
+      <NotLoggedInSection />
+    );
   }
 }
 
@@ -21,11 +29,11 @@ const LoggedInSection = () => {
     axios({
       url: backendURL + '/auth/user',
       headers: {
-        'x-auth-snailycad-token': sessionStorage.getItem('token'),
+        'x-auth-snailycad-token': Cookies.get('__session'),
       },
-    }).then(res => {
+    }).then((res) => {
       if (!res.data.user) {
-        window.location = "/auth/login"
+        window.location = '/auth/login';
       } else {
         setUsername(res.data.user[0].username);
       }
@@ -41,17 +49,19 @@ const LoggedInSection = () => {
         to='/citizen'>
         Citizens Page
       </Link>
-      <div className="mt-5">
-      </div>
+      <div className='mt-5'></div>
     </div>
   );
 };
 
 const NotLoggedInSection = () => {
-  const message = sessionStorage.getItem("home-message");
+  const message = sessionStorage.getItem('home-message');
   useEffect(() => {
-    document.addEventListener("beforeunload", sessionStorage.removeItem("home-message"))
-  })
+    document.addEventListener(
+      'beforeunload',
+      sessionStorage.removeItem('home-message')
+    );
+  });
   return (
     <div className='container mt-3 text-light'>
       {message ? <SuccessMessage message={message} dismiss /> : null}
@@ -72,12 +82,17 @@ const NotLoggedInSection = () => {
   );
 };
 
-
 const Credits = () => {
   return (
-    <div className="fixed-bottom text-light bg-dark border-dark pl-2 pt-2">
-      <p>Feel Free to support me <a href="https://www.paypal.me/caspertheghosty">here</a></p>
-      <p><a href="https://dev-caspertheghost.github.io/">CasperTheGhost</a> | Thanks for choosing SnailyCAD!</p>
+    <div className='fixed-bottom text-light bg-dark border-dark pl-2 pt-2'>
+      <p>
+        Feel Free to support me{' '}
+        <a href='https://www.paypal.me/caspertheghosty'>here</a>
+      </p>
+      <p>
+        <a href='https://dev-caspertheghost.github.io/'>CasperTheGhost</a> |
+        Thanks for choosing SnailyCAD!
+      </p>
     </div>
-  )
-}
+  );
+};