+++ server/utils/checkForUpdates.js
@@ -4,7 +4,7 @@ const chalk = require("chalk");
 
 async function checkForUpdates() {
     const result = await fetch("https://dev-caspertheghost.github.io/api/version").then(res => res.json());
-
+    
     if (result.snailycad !== package.version) {
         console.log(chalk.red.bold(`
 -------------------------------\n