+++ CHANGELOG.md
@@ -1,1 +1,12 @@
 # Changelog
+
+## 1.0.1
+
+- Fixed auto logout after closing window on home page
+- Made it a little more clear to know when registering a vehicle to a company
+- Added company addresses, you can set an address for your company.
+- Fixed `"dispatch_status" does not have a default value`
+
+## 1.0.0
+
+Initial Release